# design/API.yml
openapi: '3.0.2'
info:
  title: Cheese Community API
  version: '1.0'
servers:
  - url: http://localhost:7777/api/{version}
    description: Local server
    variables:
      version:
        default: v1
components:
  parameters:
    questionId:
      name: questionId
      in: path
      description: 问题 ID
      required: true
      schema:
        type: integer
        format: int64
    userId:
      name: userId
      in: path
      description: 用户 ID
      required: true
      schema:
        type: integer
        format: int64
    groupId:
      name: groupId
      in: path
      description: 圈子 ID
      required: true
      schema:
        type: integer
        format: int64
...
    page_start:
      name: page_start
      in: query
      description: 该页第一个 item 的 ID，留空即为从头开始
      schema:
        type: integer
        format: int64
    page_size:
      name: page_size
      in: query
      description: 每页 item 数量
      schema:
        type: integer
        format: int64
        default: 20
  securitySchemes:
    bearerAuth:
      description: The access token is used in this way.
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuth2:
      type: apiKey
      description: The refresh token is used in this way.
      name: REFRESH_TOKEN
      in: cookie
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          default: 0
          description: 错误码，0表示成功，其他表示失败
        message:
          type: string
          default: ''
          description: 错误信息，成功时为空字符串
      required:
        - code
        - message
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 用户 ID
          minimum: 1
        username:
          type: string
          description: 用户名
          example: 'cheese'
        nickname:
          type: string
          description: 昵称
          example: '芝士'
        avatar:
          type: string
          description: 头像 URL
        intro:
          type: string
          description: 个人简介
          default: 'This user has not set an introduction yet.'
        follow_count:
          type: integer
          format: int64
          description: 关注用户的数量
          minimum: 0
          example: 114
        fans_count:
          type: integer
          format: int64
          description: 粉丝数量
          minimum: 0
          example: 514
        question_count:
          type: integer
          format: int64
          description: 提问数量
          minimum: 0
          example: 1919
        answer_count:
          type: integer
          format: int64
          description: 回答数量
          minimum: 0
          example: 810
        is_follow:
          type: boolean
          description: 当前登录用户是否关注该用户
      required:
        - id
        - username
        - nickname
        - avatar
        - intro
        - follow_count
        - fans_count
        - question_count
        - answer_count
    Page:
      type: object
      description: 分页信息
      properties:
        page_start:
          type: integer
          format: int64
          description: 该页第一个 item 的 ID
        page_size:
          type: integer
          format: int64
          description: 每页 item 数量
        has_prev:
          type: boolean
          description: 是否有上一页
        prev_start:
          type: integer
          format: int64
          description: 上一页第一个 item 的 ID
        has_more:
          type: boolean
          description: 是否有下一页
        next_start:
          type: integer
          format: int64
          description: 下一页第一个 item 的 ID
    Topic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
        - id
        - name
    Question:
      type: object
      description: 问题详情
      properties:
        id:
          type: integer
          format: int64
          description: 问题 ID
        title:
          type: string
          description: 问题标题
        content:
          type: string
          description: 问题详情
        author:
          $ref: '#/components/schemas/User'
        type:
          type: integer
          description: 问题类型
        topics:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
        created_at:
          type: integer
          format: int64
          description: 创建时间（时间戳）
        updated_at:
          type: integer
          format: int64
          description: 最后修改时间（时间戳）
        is_follow:
          type: boolean
          description: 是否关注问题
        is_like:
          type: boolean
          description: 是否点赞问题
        answer_count:
          type: integer
          description: 回答数量
        comment_count:
          type: integer
          description: 评论数量
        follow_count:
          type: integer
          description: 关注数量
        like_count:
          type: integer
          description: 点赞数量
        view_count:
          type: integer
          description: 浏览数量
        is_group:
          type: boolean
          description: 是否为小组问题
        group:
          $ref: '#/components/schemas/Group'
    QuestionInvitation:
      type: object
      description: 问题邀请
      properties:
        id:
          type: integer
          format: int64
          description: 邀请 ID
        question_id:
          type: integer
          format: int64
          description: 问题 ID
        user:
          $ref: '#/components/schemas/User'
        created_at:
          type: integer
          format: int64
          description: 创建时间（时间戳）
        updated_at:
          type: integer
          format: int64
          description: 最后修改时间（时间戳）
        is_answered:
          type: boolean
          description: 是否已回答
...
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        intro:
          type: string
          default: ''
        owner:
          $ref: '#/components/schemas/User'
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
        member_count:
          type: integer
          format: int64
        question_count:
          type: integer
          format: int64
        answer_count:
          type: integer
          format: int64
        is_member:
          type: boolean
        is_owner:
          type: boolean
        is_public:
          type: boolean
      required:
        - id
        - name
        - intro
...
paths:
...
  /users/{userId}/questions:
    get:
      summary: 获取用户提问列表
      operationId: getUserQuestions
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          questions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Question'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
...
  /users/{userId}/follow/questions:
    get:
      summary: 获取用户关注的问题列表
      operationId: getUserFollowQuestions
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          questions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Question'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
...
  /questions:
    get:
      summary: 获取问题列表
      operationId: getQuestions
      parameters:
        - name: q
          in: query
          description: 搜索关键词
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          questions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Question'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
    post:
      summary: 提问
      operationId: askQuestion
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                type:
                  type: integer
                  description: 问题类型
                topics:
                  type: array
                  items:
                    type: integer
                    format: int64
                group_id:
                  type: integer
                  format: int64
                  description: 所属小组 ID，留空表示不属于任何小组
              required:
                - title
                - content
                - type
                - topics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: int64
                    required:
                      - data
  /questions/{questionId}:
    get:
      summary: 获取问题详情
      operationId: getQuestionDetail
      parameters:
        - $ref: '#/components/parameters/questionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          question:
                            $ref: '#/components/schemas/Question'
                    required:
                      - data
    put:
      summary: 更新问题
      operationId: updateQuestion
      parameters:
        - $ref: '#/components/parameters/questionId'
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                type:
                  type: integer
                  description: 问题类型
                topics:
                  type: array
                  items:
                    type: integer
                    format: int64
              required:
                - title
                - content
                - type
                - topics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
    delete:
      summary: 删除问题
      operationId: deleteQuestion
      parameters:
        - $ref: '#/components/parameters/questionId'
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
  /questions/{questionId}/followers:
    get:
      summary: 获取问题关注者列表
      operationId: getQuestionFollowers
      parameters:
        - $ref: '#/components/parameters/questionId'
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
    put:
      summary: 关注问题
      operationId: followQuestion
      parameters:
        - $ref: '#/components/parameters/questionId'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          follow_count:
                            type: integer
                            description: 新的关注者数量
                        required:
                          - follow_count
                    required:
                      - data
    delete:
      summary: 取消关注问题
      operationId: unfollowQuestion
      parameters:
        - $ref: '#/components/parameters/questionId'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          follow_count:
                            type: integer
                            description: 新的关注者数量
                        required:
                          - follow_count
                    required:
                      - data
  /questions/{id}/invitations:
    get:
      summary: 获取该问题下被邀请回答的用户
      operationId: getQuestionInvitations
      parameters:
        - name: id
          in: path
          description: 问题 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: page_start
          in: query
          description: 该页第一个用户的 ID，留空即为从头开始
          schema:
            type: integer
            format: int64
        - name: page_size
          in: query
          description: 每页用户数量
          schema:
            type: integer
            format: int64
            default: 20
        - name: sort
          in: query
          description: 排序方式字符串，例如 "+created_at" 表示按创建时间升序，"-created_at" 表示按创建时间降序，使用逗号来分隔先后多个排序字段
          schema:
            type: string
            example: '+created_at'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          invitations:
                            type: array
                            items:
                              $ref: '#/components/schemas/QuestionInvitation'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
    post:
      summary: 邀请用户回答问题
      operationId: inviteUserAnswerQuestion
      parameters:
        - name: id
          in: path
          description: 问题 ID
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  format: int64
              required:
                - user_id
      responses:
        '201':
          description: Successfully invited
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          invitation_id:
                            type: integer
                            description: 邀请 ID（如果邀请成功）
                            format: int64
                        required:
                          - invitation_id
                    required:
                      - data
...
  /questions/{id}/invitations/{invitation_id}:
    get:
      summary: 获取邀请详情
      operationId: getInvitationDetail
      parameters:
        - name: id
          in: path
          description: 问题 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: invitation_id
          in: path
          description: 邀请 ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          invitation:
                            $ref: '#/components/schemas/QuestionInvitation'
                    required:
                      - data
    delete:
      summary: 撤销邀请
      operationId: cancelInvitation
      parameters:
        - name: id
          in: path
          description: 问题 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: invitation_id
          in: path
          description: 邀请 ID
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
...
  /groups/{groupId}/questions:
    get:
      summary: 获取圈子问题列表
      operationId: getGroupQuestions
      parameters:
        - $ref: '#/components/parameters/groupId'
        - $ref: '#/components/parameters/page_start'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          questions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Question'
                          page:
                            $ref: '#/components/schemas/Page'
                    required:
                      - data
...